+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-Development:
+1.5.8:
+  - allow raw deposit transactions to be supplied to "deposit verify"
+  - move functionality of "account withdrawalcredentials" to be part of "account info"
   - add genesis validators root to "chain info"